feat(mobile): enhance 2FA settings and loading experience
- Renamed `Setting2FA` to `TwoFASettingScreen` for clarity and consistency.
- Removed the deprecated `Setting2FAScreen` component.
- Updated loading behavior to support both promise-based and immediate loading states.
- Integrated `ActivityIndicator` for iOS and retained `RotateableLoading` for other platforms.
- Improved loading overlay with a `BlurView` for better visual feedback during operations.
