+++ apps/mobile/src/lib/loading.tsx
@@ -1,6 +1,14 @@
+import { BlurView } from "expo-blur"
 import type { FC } from "react"
 import { useEffect, useRef, useState } from "react"
-import { Pressable, StyleSheet, TouchableOpacity, View } from "react-native"
+import {
+  ActivityIndicator,
+  Platform,
+  Pressable,
+  StyleSheet,
+  TouchableOpacity,
+  View,
+} from "react-native"
 import RootSiblings from "react-native-root-siblings"
 
 import { FullWindowOverlay } from "../components/common/FullWindowOverlay"
@@ -8,13 +16,20 @@ import { RotateableLoading } from "../components/common/RotateableLoading"
 import { CloseCuteReIcon } from "../icons/close_cute_re"
 
 class LoadingStatic {
-  async start<T>(promise: Promise<T>) {
+  start(): { done: () => void }
+  start<T>(promise: Promise<T>): Promise<T>
+  start<T>(promise?: Promise<T>) {
     const siblings = new RootSiblings(<LoadingContainer cancel={() => siblings.destroy()} />)
 
-    try {
-      return await promise
-    } finally {
-      siblings.destroy()
+    if (promise) {
+      promise.finally(() => siblings.destroy())
+      return promise
+    } else {
+      return {
+        done: () => {
+          siblings.destroy()
+        },
+      }
     }
   }
 }
@@ -42,13 +57,18 @@ const LoadingContainer: FC<{
     <FullWindowOverlay>
       {/* Pressable to prevent the overlay from being clicked */}
       <Pressable style={StyleSheet.absoluteFillObject} className="items-center justify-center">
-        <View className="relative rounded-2xl border border-white/20 bg-black/90 p-12">
-          <RotateableLoading />
+        <View className="border-non-opaque-separator relative overflow-hidden rounded-2xl border p-12">
+          <BlurView style={StyleSheet.absoluteFillObject} tint="systemChromeMaterialDark" />
+          {Platform.OS === "ios" ? (
+            <ActivityIndicator size={"large"} color="white" />
+          ) : (
+            <RotateableLoading />
+          )}
         </View>
         {showCancelButton && (
           <View className="absolute inset-x-0 bottom-24 flex-row justify-center">
             <TouchableOpacity onPress={cancel}>
-              <View className="rounded-full border border-white/30 p-2">
+              <View className="border-opaque-separator rounded-full border-2 p-2">
                 <CloseCuteReIcon color="gray" height={20} width={20} />
               </View>
             </TouchableOpacity>
